Security update

+++ resources/projectPhoto/photoController.js
@@ -4,6 +4,8 @@ const uuid = require('uuid/v1');
 const AWS = require('aws-sdk');
 const db = require('../../data/dbConfig');
 
+const userMatches = require('../utils/userMatches');
+
 const accessId = process.env.ACCESS_KEY_ID;
 const accessKey = process.env.SECRET_ACCESS_KEY;
 
@@ -76,8 +78,26 @@ exports.getPhotosByProjectId = async (req, res) => {
   }
 };
 
+
 exports.createProjectPhoto = async (req, res) => {
   console.log(req.body);
+
+  const project= await go.getById('user_projects', req.body.projectId);
+
+  if(project.length == 0) {
+    return res
+      .status(404)
+      .json({ message: 'A project with that ID could not be found!' });
+  }
+  if (! await userMatches(req.user, project[0].userId)) {
+    return res
+        .status(401)
+        .json({
+          message:
+            "Unauthorized: You may not add photos to this project."
+        }); 
+
+  }
   try {
     const [id] = await go.createOne('project_photos', 'id', req.body);
     res.status(201).json({ message: 'Photo successfully created', id });
@@ -85,15 +105,39 @@ exports.createProjectPhoto = async (req, res) => {
     console.error(err);
     res.status(400).json({ message: 'Unable to create photo' });
   }
+  
 };
 
 exports.deletePhotoById = async (req, res) => {
   const { id } = req.params;
+
   try {
-    await go.destroyById('project_photos', id);
-    res.json({ message: 'Successfully deleted photo' });
-  } catch (err) {
+    const data = await go.getById('project_photos', id);
+    
+    console.log("data", data);
+    
+  if (data.length === 0) {
+    res
+      .status(404)
+      .json({ message: 'A photo with that ID could not be found!' });
+  } else {
+    const project = await go.getById('user_projects', data[0].projectId);
+
+    if (await userMatches(req.user, project[0].userId)) {
+      await go.destroyById('project_photos', id);
+      res.status(200).json({ message: 'Successfully deleted photo.' });
+    } else {
+      res
+        .status(401)
+        .json({
+          message:
+            "Unauthorized: You may not delete photos that don't belong to you."
+        });
+    }
+  }   
+  }catch (err) {
     console.error(err);
     res.json({ message: 'Unable to delete photo' });
   }
 };
+
