+++ resources/users/userController.js
@@ -79,6 +79,11 @@ exports.getAllUsers = async (req, res) => {
 
 exports.updateUserById = async (req, res) => {
   const { id } = req.params;
+
+  if(!await userMatches(req.user, id)) {
+    return res.status(401).json({message: "You may not update another user's profile."});
+  }
+
   const { auth0Id, website } = req.body;
   const regex = /^(https?:\/\/)/i;
   if (!website.match(regex) && website.length > 0)
@@ -102,6 +107,14 @@ exports.updateUserById = async (req, res) => {
 
 exports.deleteUserById = async (req, res) => {
   const { id } = req.params;
+
+  if(!await userMatches(req.user, id)) {
+    console.log("No match!");
+    return res.status(401).json({message: "You may not delete another user's profile."})
+  }
+
+  console.log("yes match!");
+
   try {
     const user = await go.destroyById('users', id);
     if (user === 1) {
